+++ transaction/core/src/mint/mod.rs
@@ -4,10 +4,10 @@
 
 mod config;
 mod tx;
-mod validate;
+mod validation;
 
 pub mod constants;
 
 pub use config::{MintConfig, SetMintConfigTx, SetMintConfigTxPrefix};
 pub use tx::{MintTx, MintTxPrefix};
-pub use validate::config::{validate_set_mint_config_tx, Error as ValidateSetMintConfigTxError};
+pub use validation::{config::validate_set_mint_config_tx, error::Error as MintValidationError};
