+++ transaction/core/src/mint/validation/tx.rs
@@ -1,0 +1,33 @@
+// Copyright (c) 2018-2022 The MobileCoin Foundation
+
+//! MintTx transaction validation.
+
+/*
+use crate::{
+    mint::{
+        config::{MintTx},
+        constants::{NONCE_MAX_LENGTH, NONCE_MIN_LENGTH},
+    },
+    validation::{validate_tombstone, TransactionValidationError},
+    BlockVersion, TokenId,
+};
+use displaydoc::Display;
+use mc_crypto_keys::Ed25519Public;
+use mc_crypto_multisig::SignerSet;
+use serde::{Deserialize, Serialize};
+
+/// Determines if the transaction is valid, with respect to the provided
+/// context.
+///
+/// # Arguments
+/// * `tx` - A pending transaction.
+/// * `current_block_index` - The index of the current block that is being
+///   built.
+/// * `block_version` - The version of the block that is being built.
+pub fn validate_set_mint_config_tx(
+    tx: &MintTx,
+    current_block_index: u64,
+    block_version: BlockVersion,
+) -> Result<(), Error> {
+}
+*/
